# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestión de recolección de residuos",
        "title": "API Recolecta",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/anomalias/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anomalia"
                ],
                "summary": "Listar anomalías",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.AnomaliaListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Abierto a cualquier usuario autenticado (ciudadano, conductor\no staff) -- solo el resto del CRUD de anomalias queda\nrestringido a ADMIN/SUPERVISOR/COORDINADOR.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anomalia"
                ],
                "summary": "Crear anomalía",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.CreateAnomaliaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.AnomaliaResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/historial-asignacion/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HistorialAsignacion"
                ],
                "summary": "Obtener historial por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID historial",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.HistorialAsignacionCamionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HistorialAsignacion"
                ],
                "summary": "Actualizar historial de asignación",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID historial",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.UpdateHistorialAsignacionCamionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.HistorialAsignacionCamionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HistorialAsignacion"
                ],
                "summary": "Eliminar historial de asignación",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID historial",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.HistorialAsignacionCamionMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/ruta-camion/exists/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RutaCamion"
                ],
                "summary": "Verificar existencia ruta-camión",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.HistorialAsignacionCamionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/tipo-camion/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TipoCamion"
                ],
                "summary": "Listar tipos de camión",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.HistorialAsignacionCamionListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TipoCamion"
                ],
                "summary": "Crear tipo de camión",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.CreateHistorialAsignacionCamionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entities.HistorialAsignacionCamionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/core.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "core.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "core.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/core.ErrorDetail"
                }
            }
        },
...
        "entities.Anomalia": {
            "type": "object",
            "properties": {
                "accion_sugerida": {
                    "type": "string"
                },
                "anomalia_id": {
                    "type": "integer"
                },
                "anomalia_referencia_id": {
                    "type": "integer"
                },
                "camion_id": {
                    "type": "integer"
                },
                "categoria_clasificada": {
                    "type": "string"
                },
                "conductor_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "descripcion": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "estado": {
                    "type": "string"
                },
                "estado_pipeline": {
                    "description": "Pipeline modelo_reportes -\u003e clasificador_reportes. Se llenan por\nActualizarPipeline en background, no al crear la anomalia (por eso\nEstadoPipeline arranca en \"pendiente\" via DEFAULT de la columna).",
                    "type": "string"
                },
                "fecha_reporte": {
                    "type": "string"
                },
                "fecha_resolucion": {
                    "type": "string"
                },
                "inferencia_id": {
                    "type": "integer"
                },
                "json_ruta": {
                    "type": "string"
                },
                "nivel_riesgo": {
                    "type": "string"
                },
                "pipeline_error": {
                    "type": "string"
                },
                "punto_id": {
                    "type": "integer"
                },
                "ruta_id": {
                    "type": "integer"
                },
                "subtipo_clasificado": {
                    "type": "string"
                },
                "tipo_anomalia": {
                    "$ref": "#/definitions/entities.TipoAnomalia"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.AnomaliaListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Anomalia"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.AnomaliaResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/entities.Anomalia"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.CreateAnomaliaRequest": {
            "type": "object",
            "required": [
                "descripcion",
                "fecha_reporte",
                "tipo_anomalia"
            ],
            "properties": {
                "anomalia_referencia_id": {
                    "type": "integer"
                },
                "camion_id": {
                    "type": "integer"
                },
                "conductor_id": {
                    "type": "integer"
                },
                "descripcion": {
                    "type": "string"
                },
                "estado": {
                    "type": "string"
                },
                "fecha_reporte": {
                    "description": "Formato ISO 8601: YYYY-MM-DDTHH:MM:SSZ",
                    "type": "string",
                    "example": "2026-07-22T19:30:00Z"
                },
                "json_ruta": {
                    "type": "string"
                },
                "punto_id": {
                    "type": "integer"
                },
                "ruta_id": {
                    "type": "integer"
                },
                "tipo_anomalia": {
                    "description": "Valores validos: ANOMALIA | INCIDENCIA | REPORTE_CONDUCTOR | REPORTE_FALLA_CRITICA | SEGUIMIENTO_FALLA_CRITICA",
                    "type": "string",
                    "example": "REPORTE_CONDUCTOR"
                }
            }
        },
...
        "entities.CreateHistorialAsignacionCamionRequest": {
            "type": "object",
            "properties": {
                "fecha_asignacion": {
                    "type": "string"
                },
                "fecha_baja": {
                    "type": "string"
                },
                "id_camion": {
                    "type": "integer"
                },
                "id_chofer": {
                    "type": "integer"
                },
                "id_historial": {
                    "type": "integer"
                }
            }
        },
...
        "entities.HistorialAsignacionCamion": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "eliminado": {
                    "type": "boolean"
                },
                "fecha_asignacion": {
                    "type": "string"
                },
                "fecha_baja": {
                    "type": "string"
                },
                "id_camion": {
                    "type": "integer"
                },
                "id_chofer": {
                    "type": "integer"
                },
                "id_historial": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.HistorialAsignacionCamionListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.HistorialAsignacionCamion"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "entities.HistorialAsignacionCamionMessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "entities.HistorialAsignacionCamionResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/entities.HistorialAsignacionCamion"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entities.TipoAnomalia": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "TipoAnomaliaAnomalia",
                "TipoAnomaliaIncidencia",
                "TipoAnomaliaReporteConductor",
                "TipoAnomaliaReporteFallaCritica",
                "TipoAnomaliaSeguimientoFallaCritica"
            ]
        },
...
        "entities.UpdateHistorialAsignacionCamionRequest": {
            "type": "object",
            "properties": {
                "fecha_asignacion": {
                    "type": "string"
                },
                "fecha_baja": {
                    "type": "string"
                },
                "id_camion": {
                    "type": "integer"
                },
                "id_chofer": {
                    "type": "integer"
                },
                "id_historial": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe 'Bearer ' seguido de tu token JWT. Ejemplo: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}